fix(database): 修复数据库 count 转化 rust 类型问题

+++ server/packages/graphql/src/schema/user_info.rs
@@ -31,7 +31,7 @@ impl From<proto::user_manage::UserInfo> for UserInfo {
 #[derive(SimpleObject)]
 pub struct UserList {
     /// 总数
-    pub total: u32,
+    pub total: i64,
     /// 数据
     pub data: Vec<UserInfo>,
 }
